fix(mobile): Claude accounts move to Claude provider config + crash on add
Two issues from real-device testing:

1. Adding a Claude account asserted with "Obscured fields cannot be
   multiline" because both the create form's OAuth-token field and
   the SetClaudeTokenScreen used obscureText together with maxLines
   > 1 — Flutter forbids that combination. The OAuth blob is
   multi-line JSON that the operator pasted from a desktop login,
   so showing it (so they can verify it landed cleanly) is the
   right default. Dropped obscureText + the visibility toggle on
   both screens.

2. The Claude accounts section sat as a peer of the CLI providers
   list, which read as a flat-bug because no other provider has an
   equivalent concept. Extracted ClaudeAccountsSection into its own
   self-contained widget and embedded it inside the Claude provider
   config page, after the manifest-driven config fields. The
   section header reads "ACCOUNTS" because the surrounding context
   is already "Claude". ProvidersScreen is now a clean list of
   providers, each tapping into its own config page.
