+++ app/mobile/lib/features/providers/claude_accounts_section.dart
@@ -1,0 +1,442 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'package:opendray/core/api/api_exception.dart';
+import 'package:opendray/core/api/claude_accounts_api.dart';
+import 'package:opendray/core/api/models.dart';
+import 'package:opendray/features/providers/claude_account_dialogs.dart';
+
+// Self-contained widget that owns the Claude accounts list (fetch +
+// per-row actions). Lives inside the Claude provider's config page
+// because accounts are scoped to the Claude provider — the other
+// providers (codex / gemini / shell) have no equivalent concept, so
+// surfacing a peer "Claude accounts" section under all providers
+// reads as a flat-list bug.
+class ClaudeAccountsSection extends ConsumerStatefulWidget {
+  const ClaudeAccountsSection({super.key});
+
+  @override
+  ConsumerState<ClaudeAccountsSection> createState() =>
+      _ClaudeAccountsSectionState();
+}
+
+class _ClaudeAccountsSectionState
+    extends ConsumerState<ClaudeAccountsSection> {
+  AsyncValue<List<ClaudeAccountSummary>> _state =
+      const AsyncValue.loading();
+  final Set<String> _busy = {};
+
+  @override
+  void initState() {
+    super.initState();
+    _load();
+  }
+
+  Future<void> _load() async {
+    try {
+      final list = await ref.read(claudeAccountsApiProvider).list();
+      if (!mounted) return;
+      list.sort((a, b) {
+        if (a.isUsable != b.isUsable) return a.isUsable ? -1 : 1;
+        return a.displayName
+            .toLowerCase()
+            .compareTo(b.displayName.toLowerCase());
+      });
+      setState(() => _state = AsyncValue.data(list));
+    } on ApiException catch (e) {
+      if (mounted) {
+        setState(() => _state = AsyncValue.error(e, StackTrace.current));
+      }
+    } on Object catch (e, st) {
+      if (mounted) setState(() => _state = AsyncValue.error(e, st));
+    }
+  }
+
+  Future<void> _runOp({
+    required String key,
+    required String okMsg,
+    required String failPrefix,
+    required Future<void> Function() op,
+  }) async {
+    setState(() => _busy.add(key));
+    final messenger = ScaffoldMessenger.of(context);
+    try {
+      await op();
+      if (!mounted) return;
+      messenger.showSnackBar(
+        SnackBar(
+          content: Text(okMsg),
+          duration: const Duration(seconds: 2),
+          behavior: SnackBarBehavior.floating,
+        ),
+      );
+      await _load();
+    } on ApiException catch (e) {
+      if (!mounted) return;
+      messenger.showSnackBar(
+        SnackBar(content: Text('$failPrefix: ${e.message}')),
+      );
+    } on Object catch (e) {
+      if (!mounted) return;
+      messenger.showSnackBar(SnackBar(content: Text('$failPrefix: $e')));
+    } finally {
+      if (mounted) setState(() => _busy.remove(key));
+    }
+  }
+
+  Future<void> _add() async {
+    final form = await CreateClaudeAccountScreen.push(context);
+    if (form == null || !mounted) return;
+    await _runOp(
+      key: 'new',
+      okMsg: 'Account ${form.name} added.',
+      failPrefix: 'Add failed',
+      op: () => ref.read(claudeAccountsApiProvider).create(
+            name: form.name,
+            displayName: form.displayName,
+            token: form.token,
+          ),
+    );
+  }
+
+  Future<void> _openActions(ClaudeAccountSummary a) async {
+    final action = await showModalBottomSheet<_AccountAction>(
+      context: context,
+      backgroundColor: Theme.of(context).colorScheme.surface,
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.vertical(top: Radius.circular(16)),
+      ),
+      builder: (sheetCtx) => SafeArea(
+        top: false,
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            Padding(
+              padding: const EdgeInsets.fromLTRB(16, 14, 16, 8),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(a.displayName,
+                      style: Theme.of(sheetCtx).textTheme.titleSmall),
+                  Text(
+                    a.name,
+                    style: Theme.of(sheetCtx)
+                        .textTheme
+                        .bodySmall
+                        ?.copyWith(fontFamily: 'monospace'),
+                  ),
+                ],
+              ),
+            ),
+            const Divider(height: 1),
+            ListTile(
+              leading: Icon(
+                a.enabled ? Icons.pause_circle_outline : Icons.play_circle_outline,
+              ),
+              title: Text(a.enabled ? 'Disable' : 'Enable'),
+              onTap: () =>
+                  Navigator.of(sheetCtx).pop(_AccountAction.toggleEnabled),
+            ),
+            ListTile(
+              leading: const Icon(Icons.drive_file_rename_outline),
+              title: const Text('Rename'),
+              onTap: () => Navigator.of(sheetCtx).pop(_AccountAction.rename),
+            ),
+            ListTile(
+              leading: const Icon(Icons.key_outlined),
+              title: Text(a.tokenFilled ? 'Replace token' : 'Set token'),
+              onTap: () =>
+                  Navigator.of(sheetCtx).pop(_AccountAction.setToken),
+            ),
+            const Divider(height: 1),
+            ListTile(
+              leading: Icon(
+                Icons.delete_outline,
+                color: Theme.of(sheetCtx).colorScheme.error,
+              ),
+              title: Text(
+                'Delete',
+                style: TextStyle(color: Theme.of(sheetCtx).colorScheme.error),
+              ),
+              onTap: () => Navigator.of(sheetCtx).pop(_AccountAction.delete),
+            ),
+            const SizedBox(height: 4),
+          ],
+        ),
+      ),
+    );
+    if (action == null || !mounted) return;
+    switch (action) {
+      case _AccountAction.toggleEnabled:
+        final next = !a.enabled;
+        await _runOp(
+          key: 'a:${a.id}',
+          okMsg: next ? '${a.displayName} enabled.' : '${a.displayName} disabled.',
+          failPrefix: 'Toggle failed',
+          op: () => ref
+              .read(claudeAccountsApiProvider)
+              .setEnabled(a.id, enabled: next),
+        );
+      case _AccountAction.rename:
+        final next = await RenameClaudeAccountDialog.show(context, a);
+        if (next == null || !mounted) return;
+        await _runOp(
+          key: 'a:${a.id}',
+          okMsg: 'Renamed to $next.',
+          failPrefix: 'Rename failed',
+          op: () => ref
+              .read(claudeAccountsApiProvider)
+              .update(a.id, displayName: next),
+        );
+      case _AccountAction.setToken:
+        final tok = await SetClaudeTokenScreen.push(context, a);
+        if (tok == null || !mounted) return;
+        await _runOp(
+          key: 'a:${a.id}',
+          okMsg: 'Token saved.',
+          failPrefix: 'Set token failed',
+          op: () => ref.read(claudeAccountsApiProvider).setToken(a.id, tok),
+        );
+      case _AccountAction.delete:
+        final ok = await showDialog<bool>(
+          context: context,
+          builder: (ctx) => AlertDialog(
+            title: const Text('Delete account?'),
+            content: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  '${a.displayName} (${a.name})',
+                  style: Theme.of(ctx).textTheme.bodyMedium,
+                ),
+                const SizedBox(height: 8),
+                Text(
+                  'Removes the account and its stored OAuth token. '
+                  'Sessions already using this account stay running '
+                  'until they end.',
+                  style: Theme.of(ctx).textTheme.bodySmall,
+                ),
+              ],
+            ),
+            actions: [
+              TextButton(
+                onPressed: () => Navigator.of(ctx).pop(false),
+                child: const Text('Cancel'),
+              ),
+              FilledButton(
+                style: FilledButton.styleFrom(
+                  backgroundColor: Theme.of(ctx).colorScheme.error,
+                ),
+                onPressed: () => Navigator.of(ctx).pop(true),
+                child: const Text('Delete'),
+              ),
+            ],
+          ),
+        );
+        if (ok != true || !mounted) return;
+        await _runOp(
+          key: 'a:${a.id}',
+          okMsg: 'Deleted ${a.name}.',
+          failPrefix: 'Delete failed',
+          op: () => ref.read(claudeAccountsApiProvider).delete(a.id),
+        );
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final muted = Theme.of(context).textTheme.bodySmall;
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        Padding(
+          padding: const EdgeInsets.fromLTRB(4, 14, 0, 6),
+          child: Row(
+            children: [
+              Expanded(
+                child: Text(
+                  'ACCOUNTS',
+                  style: Theme.of(context).textTheme.labelSmall?.copyWith(
+                        letterSpacing: 0.8,
+                        color: Theme.of(context)
+                            .colorScheme
+                            .onSurface
+                            .withValues(alpha: 0.6),
+                      ),
+                ),
+              ),
+              TextButton.icon(
+                icon: const Icon(Icons.add, size: 16),
+                label: const Text('Add'),
+                onPressed: _add,
+              ),
+            ],
+          ),
+        ),
+        _state.when(
+          data: _buildList,
+          loading: () => const Padding(
+            padding: EdgeInsets.symmetric(vertical: 16),
+            child: Center(
+              child: SizedBox(
+                width: 18,
+                height: 18,
+                child: CircularProgressIndicator(strokeWidth: 2),
+              ),
+            ),
+          ),
+          error: (e, _) => Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Failed to load accounts: $e',
+                  style: TextStyle(
+                    color: Theme.of(context).colorScheme.error,
+                    fontSize: 12,
+                  ),
+                ),
+                const SizedBox(height: 6),
+                OutlinedButton(
+                  onPressed: _load,
+                  child: const Text('Retry'),
+                ),
+              ],
+            ),
+          ),
+        ),
+        Padding(
+          padding: const EdgeInsets.only(top: 6, left: 4),
+          child: Text(
+            'Sessions spawned with the Claude provider pick from these '
+            'accounts. Disabled / token-empty rows are filtered out of '
+            'the picker.',
+            style: muted,
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildList(List<ClaudeAccountSummary> list) {
+    if (list.isEmpty) {
+      return Padding(
+        padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 4),
+        child: Text(
+          'No Claude accounts yet. Tap Add to paste an OAuth token '
+          'from a desktop login.',
+          style: Theme.of(context).textTheme.bodySmall,
+        ),
+      );
+    }
+    return Column(
+      children: [
+        for (final a in list)
+          _AccountTile(
+            account: a,
+            busy: _busy.contains('a:${a.id}'),
+            onTap: () => _openActions(a),
+          ),
+      ],
+    );
+  }
+}
+
+enum _AccountAction { toggleEnabled, rename, setToken, delete }
+
+class _AccountTile extends StatelessWidget {
+  const _AccountTile({
+    required this.account,
+    required this.busy,
+    required this.onTap,
+  });
+
+  final ClaudeAccountSummary account;
+  final bool busy;
+  final VoidCallback onTap;
+
+  @override
+  Widget build(BuildContext context) {
+    final usable = account.isUsable;
+    return ListTile(
+      onTap: busy ? null : onTap,
+      contentPadding: EdgeInsets.zero,
+      leading: Icon(
+        usable ? Icons.account_circle : Icons.account_circle_outlined,
+        color: usable
+            ? Theme.of(context).colorScheme.primary
+            : Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.4),
+      ),
+      title: Row(
+        children: [
+          Flexible(
+            child: Text(
+              account.displayName,
+              overflow: TextOverflow.ellipsis,
+              style: TextStyle(
+                color: account.enabled
+                    ? null
+                    : Theme.of(context)
+                        .colorScheme
+                        .onSurface
+                        .withValues(alpha: 0.55),
+              ),
+            ),
+          ),
+          const SizedBox(width: 6),
+          if (!account.enabled)
+            _MiniBadge(
+              label: 'disabled',
+              color: Theme.of(context).colorScheme.error,
+            ),
+          if (!account.tokenFilled) ...[
+            const SizedBox(width: 4),
+            _MiniBadge(
+              label: 'no token',
+              color: Theme.of(context).colorScheme.error,
+            ),
+          ],
+        ],
+      ),
+      subtitle: Text(
+        account.id,
+        style: Theme.of(context)
+            .textTheme
+            .bodySmall
+            ?.copyWith(fontFamily: 'monospace'),
+      ),
+      trailing: busy
+          ? const SizedBox(
+              width: 18,
+              height: 18,
+              child: CircularProgressIndicator(strokeWidth: 2),
+            )
+          : const Icon(Icons.more_vert),
+    );
+  }
+}
+
+class _MiniBadge extends StatelessWidget {
+  const _MiniBadge({required this.label, required this.color});
+  final String label;
+  final Color color;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 1),
+      decoration: BoxDecoration(
+        color: color.withValues(alpha: 0.15),
+        borderRadius: BorderRadius.circular(4),
+        border: Border.all(color: color.withValues(alpha: 0.6), width: 0.5),
+      ),
+      child: Text(
+        label,
+        style: TextStyle(color: color, fontSize: 10),
+      ),
+    );
+  }
+}
